+++ src/applyShellEdit.ts
@@ -1,10 +1,9 @@
 import { ExecFileOptions } from 'child_process'
-import { EOL } from 'os'
 import * as path from 'path'
 import * as util from 'util'
 import { Range, TextDocument, window } from 'vscode'
 import { getExtConfig } from './config'
-import { execFilePm, isWin32 } from './util'
+import { execFilePm, isWin32, noop } from './util'
 
 const nodeLangIds = [
   'javascript',
@@ -16,32 +15,6 @@ const nodeLangIds = [
   'html',
   'svelte',
 ]
-
-export async function applyShellEdit() {
-  const editor = window.activeTextEditor
-  if (!editor) {
-    return
-  }
-
-  const { document } = editor
-  const selectMap = new Map<Range, string>()
-
-  for (const selectionRange of editor.selections) {
-    const line = document.lineAt(selectionRange.start.line)
-    const range = selectionRange.isEmpty ? line.range : selectionRange
-    const text = selectionRange.isEmpty
-      ? line.text
-      : document.getText(selectionRange)
-    const result = await execByLangId(text, document)
-    result && selectMap.set(range, result)
-  }
-
-  await editor.edit((edit) => {
-    for (const [selectionRange, result] of selectMap) {
-      edit.replace(selectionRange, result)
-    }
-  })
-}
 
 export async function execByLangId(text: string, document: TextDocument) {
   const { languageId } = document
@@ -59,14 +32,48 @@ export async function execByLangId(text: string, document: TextDocument) {
   const options: ExecFileOptions = {
     cwd: path.dirname(document.fileName),
   }
-  return (
-    await execFilePm(cmd.shift()!, cmd, options).catch((err) => ({
-      stdout: String(err),
-    }))
-  ).stdout.trimEnd()
+  return execFilePm(cmd.shift()!, cmd, options).then(
+    (r) => r.stdout,
+    (err) => String(err),
+  )
 }
 
-export async function applyCurrentShellEdit() {
+export async function applyShellEdit() {
+  const editor = window.activeTextEditor
+  if (!editor) {
+    return
+  }
+
+  const { document, selections } = editor
+
+  const results = await Promise.all(
+    (function* () {
+      for (const selectionRange of selections) {
+        if (selectionRange.isEmpty) {
+          yield execByLangId(
+            document.lineAt(selectionRange.start.line).text,
+            document,
+          )
+        } else {
+          yield execByLangId(document.getText(selectionRange), document)
+        }
+      }
+    })(),
+  )
+
+  await editor.edit((edit) => {
+    // ignore empty results, modify it directly
+    for (let i = 0; i < selections.length; i++) {
+      if (selections[i].isEmpty) {
+        edit.replace(document.lineAt(selections[i].start).range, results[i])
+      } else {
+        edit.replace(selections[i], results[i])
+      }
+    }
+  })
+}
+
+export async function applyTerminalEdit() {
   const editor = window.activeTextEditor
   const terminal = window.activeTerminal
   if (!(editor && terminal)) {
@@ -76,28 +83,30 @@ export async function applyCurrentShellEdit() {
   const { document } = editor
   const selectMap = new Map<Range, string>()
 
-  for (const selectionRange of editor.selections) {
-    const line = document.lineAt(selectionRange.start.line)
-    const range = selectionRange.isEmpty ? line.range : selectionRange
-    const text = selectionRange.isEmpty
-      ? line.text
-      : document.getText(selectionRange)
+  for (let selectionRange of editor.selections as readonly Range[]) {
+    let text: string
+    if (selectionRange.isEmpty) {
+      const line = document.lineAt(selectionRange.start.line)
+      selectionRange = line.range
+      text = line.text
+    } else {
+      text = document.getText(selectionRange)
+    }
 
     await new Promise<void>((resolve, reject) => {
       const event = window.onDidExecuteTerminalCommand((e) => {
         if (e.terminal === terminal) {
           event.dispose()
-          const result = e.output?.trimEnd()
-          if (result) {
-            selectMap.set(range, result)
+          if (e.exitCode === 0 && e.output) {
+            selectMap.set(selectionRange, e.output)
             resolve()
           } else {
-            reject('snippets produces no outputs')
+            reject('command failed or no outputs')
           }
         }
       })
-      terminal.sendText(text)
-    })
+      terminal.sendText(text, true)
+    }).catch(noop)
   }
 
   await editor.edit((edit) => {
@@ -107,12 +116,7 @@ export async function applyCurrentShellEdit() {
   })
 }
 
-const shellFilterPrefix = {
-  pwsh: "@'\n%s\n'@ | ".replaceAll('\n', EOL),
-  bash: 'cat << EOF\n%s\nEOF | '.replaceAll('\n', EOL),
-}
-
-export async function applyShellFilter() {
+export async function applyTerminalFilter() {
   const editor = window.activeTextEditor
   const terminal = window.activeTerminal
   if (!(editor && terminal)) {
@@ -121,37 +125,48 @@ export async function applyShellFilter() {
   terminal.show()
 
   const { document, selections } = editor
-  const docEOL = ['\n', '\r\n'][document.eol - 1]
-  let text = ''
+  const lines = []
   for (const selectionRange of selections) {
     if (selectionRange.isEmpty) {
-      text += document.lineAt(selectionRange.start).text.trimEnd() + docEOL
+      lines.push(document.lineAt(selectionRange.start).text)
     } else {
-      text += document.getText(selectionRange).trimEnd() + docEOL
+      lines.push(document.getText(selectionRange).replace(/\r\n?/g, '\n'))
     }
   }
-  if (docEOL !== EOL) {
-    text = text.replaceAll(docEOL, EOL)
-  }
-  let shellType: 'pwsh' | 'bash' = isWin32 ? 'pwsh' : 'bash'
-  if (/pwsh|powershell/i.test(terminal.creationOptions.name ?? '')) {
-    shellType = 'pwsh'
-  }
-  text = util.format(shellFilterPrefix[shellType], text)
 
-  const result = await new Promise<string | undefined>((resolve) => {
+  let text = lines.join('\n')
+  let shellType: 'pwsh' | 'bash'
+  if (
+    terminal.creationOptions.name &&
+    /pwsh|powershell/i.test(terminal.creationOptions.name)
+  ) {
+    shellType = 'pwsh'
+  } else {
+    shellType = isWin32 ? 'pwsh' : 'bash'
+  }
+
+  text = util.format(
+    {
+      pwsh: "@'\n%s\n'@ | ",
+      bash: "cat << 'EOF'\n%s\nEOF | ",
+    }[shellType],
+    text,
+  )
+
+  return new Promise<string>((resolve, reject) => {
     const event = window.onDidExecuteTerminalCommand((e) => {
       if (e.terminal === terminal) {
-        resolve(e.output)
         event.dispose()
+        if (e.exitCode === 0 && e.output) {
+          resolve(e.output)
+        } else {
+          reject('command failed or no outputs')
+        }
       }
     })
     terminal.sendText(text)
-  })
-
-  if (result) {
-    await editor.edit((edit) => {
-      edit.replace(selections[0], result)
-    })
-  }
+  }).then(
+    (result) => editor.edit((edit) => edit.replace(selections[0], result)),
+    noop,
+  )
 }