+++ src/extension.ts
@@ -1,34 +1,42 @@
 import { ExtensionContext, commands } from 'vscode'
 import {
-  applyCurrentShellEdit,
   applyShellEdit,
-  applyShellFilter,
+  applyTerminalEdit,
+  applyTerminalFilter,
 } from './applyShellEdit'
 import { setExtContext } from './context'
-import { register as registerCssSeletorComplete } from './cssSelectorComplete'
+import { CssSelectorCompleteProvider } from './cssSelectorComplete'
 import { PathCompleteProvider } from './pathComplete'
 import { quicklySwitchFile } from './quicklySwitchFile'
-import { renameWithCase, transformCase } from './transformCase'
-import { register as registerTsCodeAction } from './tsCodeActions/selection'
+import {
+  renameWithCase,
+  transformCase,
+  transformCaseWithPicker,
+} from './transformCase'
+import { SelectionCodeActionsProvider } from './tsCodeActions/selection'
 
 export function activate(context: ExtensionContext) {
   setExtContext(context)
-  registerCssSeletorComplete()
-  registerTsCodeAction()
   context.subscriptions.push(
     new PathCompleteProvider(),
+    new SelectionCodeActionsProvider(),
+    new CssSelectorCompleteProvider(),
     commands.registerCommand('mvext.renameWithCase', renameWithCase),
     commands.registerTextEditorCommand('mvext.transformCase', transformCase),
-    commands.registerCommand('mvext.quicklySwitchFile', quicklySwitchFile),
+    commands.registerCommand(
+      'mvext.transformCaseWithPicker',
+      transformCaseWithPicker,
+    ),
     commands.registerCommand('mvext.applyShellEdit', applyShellEdit),
+    commands.registerCommand('mvext.quicklySwitchFile', quicklySwitchFile),
   )
   if (__DEV__) {
     context.subscriptions.push(
+      commands.registerCommand('mvext.applyTerminalEdit', applyTerminalEdit),
       commands.registerCommand(
-        'mvext.applyCurrentShellEdit',
-        applyCurrentShellEdit,
+        'mvext.applyTerminalFilter',
+        applyTerminalFilter,
       ),
-      commands.registerCommand('mvext.applyShellFilter', applyShellFilter),
     )
   }
 }